requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-docx==1.1.0
markdown==3.5.1
reportlab==4.0.7
weasyprint==61.2
ollama==0.1.6
httpx==0.25.2
aiofiles==23.2.1
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2
